Переход на сборку Maven
