--- .idea/libraries/__java.xml
@@ -1,10 +1,0 @@
-<component name="libraryTable">
-  <library name="Библиотеки для java">
-    <CLASSES>
-      <root url="file://$USER_HOME$/Desktop/Коля/Библиотеки для java" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-    <jarDirectory url="file://$USER_HOME$/Desktop/Коля/Библиотеки для java" recursive="false" />
-  </library>
-</component>
