--- .idea/libraries/commons_io_2_6.xml
@@ -1,9 +1,0 @@
-<component name="libraryTable">
-  <library name="commons-io-2.6">
-    <CLASSES>
-      <root url="jar://$USER_HOME$/Desktop/Коля/Библиотеки для java/commons-io-2.6.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
